+++ src/ai-bundle/config/services.php
@@ -169,6 +169,8 @@ return static function (ContainerConfigurator $container): void {
             ->args([
                 tagged_iterator('ai.traceable_platform'),
                 tagged_iterator('ai.traceable_toolbox'),
+                tagged_iterator('ai.traceable_message_store'),
+                tagged_iterator('ai.traceable_chat'),
             ])
             ->tag('data_collector')